pyproject.toml

[project]
name = "optimus_portfolio_api"
version = "0.1.0"
description = ""
authors = [
    {name = "Jayden", email = "jpylesbusiness@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "spotipy>=2.23.0",
    "pyyaml>=6.0.1",
    "notion-client>=2.2.1",
    "pymongo>=4.6.2",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.isort]
length_sort = "1"
profile = "black"
sections = "STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
import_heading_stdlib = "STL"
import_heading_thirdparty = "PDM"
import_heading_firstparty = "LOCAL"
import_heading_localfolder = "LOCAL"
